[FX-148] players mustnt be blocked anymore

+++ src/game/GameObject.cpp
@@ -1476,7 +1476,7 @@ void GameObject::TakenDamage(uint32 damage)
     else
         m_goValue->building.health = 0;
 
-	uint8 life = (uint8)ceil(m_goValue->building.health / (m_goInfo->building.intactNumHits + m_goInfo->building.damagedNumHits) * 255/100);
+	uint8 life = (uint8)ceil((double)m_goValue->building.health / (m_goInfo->building.intactNumHits + m_goInfo->building.damagedNumHits) * 255/100);
 	SetGoAnimProgress(life);
     if (HasFlag(GAMEOBJECT_FLAGS, GO_FLAG_DAMAGED)) // from damaged to destroyed
     {
